Add more tests.
I need to figure out why tests with stubs are breaking. There's
something going on with reseting. Busted probably avoided this problem
by reloading the world.
